Move field_tests contents to gameplay_tests
