+++ src/software/gameplay_tests/movement_robot_field_test.py
@@ -1,5 +1,5 @@
 from proto.import_all_protos import *
-from software.field_tests.field_test_fixture import *
+from software.gameplay_tests.field_test_fixture import *
 
 from software.gameplay_tests.simulated_test_fixture import *
 from software.logger.logger import create_logger
