+++ src/software/gameplay_tests/passing_field_test.py
@@ -1,6 +1,6 @@
 import software.python_bindings as tbots_cpp
 from proto.import_all_protos import *
-from software.field_tests.field_test_fixture import *
+from software.gameplay_tests.field_test_fixture import *
 from software.gameplay_tests.validation.friendly_receives_ball_slow import (
     FriendlyAlwaysReceivesBallSlow,
 )
