+++ src/software/gameplay_tests/pivot_kick_field_test.py
@@ -1,6 +1,6 @@
 import math
 from proto.import_all_protos import *
-from software.field_tests.field_test_fixture import *
+from software.gameplay_tests.field_test_fixture import *
 
 from software.gameplay_tests.simulated_test_fixture import *
 from software.logger.logger import create_logger
